// src/features/users/components/user-drawer/user-details-tab.schema.ts
import { z } from 'zod';

export const editUserProfileSchema = z.object({
  firstName: z.string().min(1, 'First name is required'),
  lastName: z.string().min(1, 'Last name is required'),
  email: z.string().email('Invalid email address'),
  phone: z.string().optional(),
});

export type EditUserProfileFormData = z.infer<typeof editUserProfileSchema>;